WIP: Move more DB operations to model
