+++ server/model/status_page.js
@@ -5,6 +5,9 @@ const { UptimeKumaServer } = require("../uptime-kuma-server");
 const { log } = require("../../src/util");
 const jsesc = require("jsesc");
 const googleAnalytics = require("../google-analytics");
+const ImageDataURI = require("../image-data-uri");
+const Database = require("../database");
+const { setSetting } = require("../util-server");
 
 class StatusPage extends BeanModel {
 
@@ -13,6 +16,187 @@ class StatusPage extends BeanModel {
      * @type {{}}
      */
     static domainMappingList = { };
+
+    /**
+     * Create new StatusPage
+     * @param {string} title title of the status page
+     * @param {string} slug slug of the satus page
+     * @returns {Promise<Bean>} The created status page
+     */
+    static async create(title, slug) {
+        // Check empty
+        if (!title || !slug) {
+            throw new Error("Please input all fields");
+        }
+
+        // Make sure slug is string
+        if (typeof slug !== "string") {
+            throw new Error("Slug -Accept string only");
+        }
+
+        // lower case only
+        slug = slug.toLowerCase();
+
+        StatusPage.checkSlug(slug);
+
+        let statusPage = R.dispense("status_page");
+        statusPage.slug = slug;
+        statusPage.title = title;
+        statusPage.theme = "auto";
+        statusPage.icon = "";
+        await R.store(statusPage);
+
+        return statusPage;
+    }
+
+    /**
+     * Saves and updates Status Page
+     *
+     * @param {string} slug slug of the status page to save to
+     * @param {Object} config config of the status page
+     * @param {string} imgDataUrl url of the logo
+     * @param {Object[]} publicGroupList list of groups of monitors
+     * @return {Promise<Bean>} the updated status page bean
+     */
+    static async save(slug, config, imgDataUrl, publicGroupList) {
+        // Save Config
+        let statusPage = await R.findOne("status_page", " slug = ? ", [
+            slug
+        ]);
+
+        if (!statusPage) {
+            throw new Error("No slug?");
+        }
+
+        StatusPage.checkSlug(config.slug);
+
+        const header = "data:image/png;base64,";
+
+        // Check logo format
+        // If is image data url, convert to png file
+        // Else assume it is a url, nothing to do
+        if (imgDataUrl.startsWith("data:")) {
+            if (! imgDataUrl.startsWith(header)) {
+                throw new Error("Only allowed PNG logo.");
+            }
+
+            const filename = `logo${statusPage.id}.png`;
+
+            // Convert to file
+            await ImageDataURI.outputFile(imgDataUrl, Database.uploadDir + filename);
+            config.logo = `/upload/${filename}?t=` + Date.now();
+
+        } else {
+            config.icon = imgDataUrl;
+        }
+
+        statusPage.slug = config.slug;
+        statusPage.title = config.title;
+        statusPage.description = config.description;
+        statusPage.icon = config.logo;
+        statusPage.theme = config.theme;
+        //statusPage.published = ;
+        //statusPage.search_engine_index = ;
+        statusPage.show_tags = config.showTags;
+        //statusPage.password = null;
+        statusPage.footer_text = config.footerText;
+        statusPage.custom_css = config.customCSS;
+        statusPage.show_powered_by = config.showPoweredBy;
+        statusPage.modified_date = R.isoDateTime();
+        statusPage.google_analytics_tag_id = config.googleAnalyticsId;
+
+        await R.store(statusPage);
+
+        await statusPage.updateDomainNameList(config.domainNameList);
+        await StatusPage.loadDomainMappingList();
+
+        // Save Public Group List
+        const groupIDList = [];
+        let groupOrder = 1;
+
+        for (let group of publicGroupList) {
+            let groupBean;
+            if (group.id) {
+                groupBean = await R.findOne("group", " id = ? AND public = 1 AND status_page_id = ? ", [
+                    group.id,
+                    statusPage.id
+                ]);
+            } else {
+                groupBean = R.dispense("group");
+            }
+
+            groupBean.status_page_id = statusPage.id;
+            groupBean.name = group.name;
+            groupBean.public = true;
+            groupBean.weight = groupOrder++;
+
+            await R.store(groupBean);
+
+            await R.exec("DELETE FROM monitor_group WHERE group_id = ? ", [
+                groupBean.id
+            ]);
+
+            let monitorOrder = 1;
+
+            for (let monitor of group.monitorList) {
+                let relationBean = R.dispense("monitor_group");
+                relationBean.weight = monitorOrder++;
+                relationBean.group_id = groupBean.id;
+                relationBean.monitor_id = monitor.id;
+
+                if (monitor.sendUrl !== undefined) {
+                    relationBean.send_url = monitor.sendUrl;
+                }
+
+                await R.store(relationBean);
+            }
+
+            groupIDList.push(groupBean.id);
+            group.id = groupBean.id;
+        }
+
+        // Delete groups that are not in the list
+        log.debug("socket", "Delete groups that are not in the list");
+        const slots = groupIDList.map(() => "?").join(",");
+
+        const data = [
+            ...groupIDList,
+            statusPage.id
+        ];
+        await R.exec(`DELETE FROM \`group\` WHERE id NOT IN (${slots}) AND status_page_id = ?`, data);
+
+        const server = UptimeKumaServer.getInstance();
+
+        // Also change entry page to new slug if it is the default one, and slug is changed.
+        if (server.entryPage === "statusPage-" + slug && statusPage.slug !== slug) {
+            server.entryPage = "statusPage-" + statusPage.slug;
+            await setSetting("entryPage", server.entryPage, "general");
+        }
+
+        return statusPage;
+    }
+
+    /**
+     * Check slug a-z, 0-9, - only
+     * Regex from: https://stackoverflow.com/questions/22454258/js-regex-string-validation-for-slug
+     * @param {string} slug Slug to test
+     * @throws if the slug is invalid
+     */
+    static checkSlug(slug) {
+        if (typeof slug !== "string") {
+            throw new Error("Slug must be string");
+        }
+
+        slug = slug.trim();
+
+        if (!slug) {
+            throw new Error("Slug cannot be empty");
+        }
+
+        if (!slug.match(/^[A-Za-z0-9]+(?:-[A-Za-z0-9]+)*$/)) {
+            throw new Error("Invalid Slug");
+        }
+    }
 
     /**
      *
@@ -120,6 +304,35 @@ class StatusPage extends BeanModel {
             config: await statusPage.toPublicJSON(),
             incident,
             publicGroupList,
+            maintenanceList,
+        };
+    }
+
+    /**
+     * Produce full data of status page for backup
+     * @param {StatusPage} statusPage
+     */
+    static async getStatusPageBackup(statusPage) {
+        // Incidents
+        let incidents = await Promise.all((await R.find("incident", "status_page_id = ?", [
+            statusPage.id,
+        ])).map(async (incident) => await incident.toPublicJSON()));
+
+        // Maintenance
+        let maintenanceList = await StatusPage.getMaintenanceList(statusPage.id);
+
+        // Monitors
+        const groupList = await Promise.all((await R.find("group", "status_page_id = ? ORDER BY weight ", [
+            statusPage.id
+        ])).map(async (group) => {
+            return await group.toPublicJSON(false);
+        }));
+
+        // Response
+        return {
+            config: await statusPage.toJSON(),
+            incidents,
+            groupList,
             maintenanceList,
         };
     }
